tests: consolidate local_stack helpers under compute_space/tests
Merge compute_space/local_stack.py and tests/local_stack.py into
compute_space/tests/local_stack.py (same shared-test-utils location as
utils.py), with make_router_env in utils.py next to managed_router.
Drop the SECRET_KEY test env var: it fed Quart's cookie-session signing
in the original implementation and nothing reads it since the Litestar
rewrite.
